Some experiments with arc fitting
There's more to be done, but checking in what I have. That includes very careful calculation of the area of an arc segment, and inverting that.

Also some explorations of area of an Euler spiral parallel curve. The result of that exploration is that fine-tuning area is not profitable, as it scales the same as the error of the underlying Euler spiral to arc approximation. A far better approach is to apply a small modifier to the error metric. The current `arc_main` contains a by-hand tester to fit that metric.