First cut at CPU performance measurement
This just compares against Skia, which is not especially informative.

Skia has a bug that causes over-subdivision at fine tolerances. Our code could also be more optimized; particularly low-hanging fruit includes not evaluating the Euler spiral twice for both sides of a stroke.

The SVG parser now takes into account the scale, which is especially important for the waves example.