Redo timings and primitive count measurements
The script for gathering timings had a bug, where the times were cumulative over the scenes. Now fixed. We also include Nehab in the timings.

As part of the work to add Nehab, there's a new utility to convert scenes into RVG format. It's crude (no colors, viewport is hardcoded), but sufficient for measuring timing and primitive count.

The primitive count measurements were redone, striking waves (as it triggers a Skia bug) and adding mmark-70k.