Rewrite of introduction
The introduction is rewritten, both to address reviewer feedback and also to strenghten our statement of strong correctness.

Also redo the transition from circular arc to the sqrt-kappa error metric, making the statement stronger, and also including the monotonic curvature condition.

More is to be done, basically making things flow more smoothly in subsequent sections, but I wanted to get this part up for review, as other revision will depend on it.