More formatting fixes
Use acmauthoryear cite style. Add email addresses. Squash linespace of algorithm 2 to fit page margins.

With this cite style, the `\citenum` command is no longer valid, so those are all changed to `\cite`. There are a number of instances that use `\citet` with the author specified in the text, and I believe those are compliant, but I'm not sure that's best. It's possible those should be edited to `\cite` so the author is inside the brackets.